Add a simulated backend to demonstrate the end-to-end integration process (#9)
This adds some basic network calls to the demo so that it's more
representative of a real app. They're now displayed alongside the client
SDK data in order to provide more usage context.

>[!CAUTION]
> This is NOT the way SnapAuth is normally used - you should never make
the server calls from client code. It will result in leaking secrets
that can lead to account takeover.

This doesn't _have_ an actual backend, and the demo app is being run on
a dedicated domain with a dedicated SnapAuth account so there's nothing
to compromise.